JIRA:GB-1978 orgmod description needed as static function for defline tools
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@59777 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/edit/autodef_feature_clause_base.hpp
@@ -55,7 +55,7 @@ public:
     string PrintClause(bool print_typeword, bool typeword_is_plural);
     
     virtual CSeqFeatData::ESubtype  GetMainFeatureSubtype();
-    unsigned int GetNumSubclauses() { return m_ClauseList.size(); }
+    size_t GetNumSubclauses() { return m_ClauseList.size(); }
     virtual void Label();
     virtual bool AddmRNA (CAutoDefFeatureClause_Base *mRNAClause);
     virtual bool AddGene (CAutoDefFeatureClause_Base *gene_clause);
